Build a separated folder + separated button
